// app/src/generated/types/Config.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from "@metaplex-foundation/beet";
export type Config = {
    price: beet.bignum;
    amountPerMint: beet.bignum;
    startDate: beet.bignum;
    counter: beet.bignum;
    supply: beet.bignum;
    tick: string;
    liquidityBootstrapping: boolean;
    filter: string[];
};

/**
 * @category userTypes
 * @category generated
 */
export const configBeet = new beet.FixableBeetArgsStruct<Config>(
    [
        ["price", beet.u64],
        ["amountPerMint", beet.u64],
        ["startDate", beet.u64],
        ["counter", beet.u64],
        ["supply", beet.u64],
        ["tick", beet.utf8String],
        ["liquidityBootstrapping", beet.bool],
        ["filter", beet.array(beet.utf8String)],
    ],
    "Config",
);